Add C-q to describe elisp function when in counsel-M-x
As this is mapped to counsel-describe-map, this command is available
in other buffers where the describe-function would have no meaning (i.e.
would give an error). I see no counsel-M-x-map on which to bind this
key on the other hand.
It might be better/cleaner to just use the actions for the counsel-M-x,
which I've also added, but that's an extra keystroke (M-o d instead of
C-q)